pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "md-converter"
version = "0.1.0"
description = "Convert web pages and documents to clean markdown for LLMs"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "flask>=3.0",
    "gunicorn>=22.0",
    "markdownify>=0.14",
    "filetype>=1.2",
    "trafilatura>=2.0",
    "playwright>=1.40",
    "pymupdf4llm>=0.0.10",
    "mammoth>=1.8",
    "python-pptx>=1.0",
    "openpyxl>=3.1",
    "pytesseract>=0.3",
    "Pillow>=10.0",
    "ebooklib>=0.18",
    "faster-whisper>=1.0",
    "defusedxml>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]
